feat: add several UI demo & common comps
