+++ .eslintrc.json
@@ -46,6 +46,6 @@
     "quotes": ["warn", "single"],
     "no-duplicate-imports": "error",
     "@typescript-eslint/no-unused-vars": "warn",
-    "tailwindcss/no-custom-classname": ["warn", {"whitelist": ["line-clamp-[1-9]{1}[0-9]?"]}]
+    "tailwindcss/no-custom-classname": ["warn", {"whitelist": ["line-clamp-[1-9]{1}[0-9]?", "card-primary", "flex-center-between"]}]
   }
 }
