+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@emotion/react": "^11.10.5",
+    "@mantine/carousel": "^5.9.5",
     "@mantine/core": "^5.8.4",
     "@mantine/dates": "^5.8.4",
     "@mantine/form": "^5.8.4",
@@ -24,6 +25,8 @@
     "@tanstack/react-query": "^4.20.2",
     "axios": "^1.2.0",
     "dayjs": "^1.11.6",
+    "embla-carousel-autoplay": "^7.0.5",
+    "embla-carousel-react": "^7.0.5",
     "i18next": "21.9.1",
     "i18next-browser-languagedetector": "6.1.5",
     "i18next-http-backend": "1.4.1",
