--- src/features/Demos/Table.tsx
@@ -1,151 +1,0 @@
-import {
-  Avatar,
-  Badge,
-  Table,
-  Group,
-  Text,
-  Select,
-  ScrollArea,
-  useMantineTheme,
-  ActionIcon,
-  Menu,
-} from '@mantine/core';
-import {
-  IconPencil,
-  IconDots,
-  IconMessages,
-  IconNote,
-  IconReportAnalytics,
-  IconTrash,
-} from '@tabler/icons';
-
-interface UsersTableProps {
-  data?: {avatar: string; name: string; job: string; email: string; role: string}[];
-}
-
-const rolesData = ['Manager', 'Collaborator', 'Contractor'];
-
-const mockDatas = [
-  {
-    avatar:
-      'https://images.unsplash.com/photo-1624298357597-fd92dfbec01d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=250&q=80',
-    name: 'Robert Wolfkisser',
-    job: 'Engineer',
-    email: 'rob_wolf@gmail.com',
-    role: 'Collaborator',
-  },
-  {
-    avatar:
-      'https://images.unsplash.com/photo-1586297135537-94bc9ba060aa?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=250&q=80',
-    name: 'Jill Jailbreaker',
-    job: 'Engineer',
-    email: 'jj@breaker.com',
-    role: 'Collaborator',
-  },
-  {
-    avatar:
-      'https://images.unsplash.com/photo-1632922267756-9b71242b1592?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=250&q=80',
-    name: 'Henry Silkeater',
-    job: 'Designer',
-    email: 'henry@silkeater.io',
-    role: 'Contractor',
-  },
-  {
-    avatar:
-      'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=250&q=80',
-    name: 'Bill Horsefighter',
-    job: 'Designer',
-    email: 'bhorsefighter@gmail.com',
-    role: 'Contractor',
-  },
-  {
-    avatar:
-      'https://images.unsplash.com/photo-1630841539293-bd20634c5d72?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=250&q=80',
-    name: 'Jeremy Footviewer',
-    job: 'Manager',
-    email: 'jeremy@foot.dev',
-    role: 'Manager',
-  },
-];
-const jobColors: Record<string, string> = {
-  engineer: 'blue',
-  manager: 'cyan',
-  designer: 'pink',
-};
-
-export const TableDemo = ({data}: UsersTableProps) => {
-  const theme = useMantineTheme();
-
-  const rows = (data || mockDatas).map((item) => (
-    <tr key={item.name}>
-      <td>
-        <Group spacing="sm">
-          <Avatar size={40} src={item.avatar} radius={40} />
-          <div>
-            <Text size="sm" weight={500}>
-              {item.name}
-            </Text>
-            <Text size="xs" color="dimmed">
-              {item.email}
-            </Text>
-          </div>
-        </Group>
-      </td>
-
-      <td>
-        <Select data={rolesData} defaultValue={item.role} variant="unstyled" />
-      </td>
-
-      <td>
-        <Badge
-          color={jobColors[item.job.toLowerCase()]}
-          variant={theme.colorScheme === 'dark' ? 'light' : 'outline'}
-        >
-          {item.job}
-        </Badge>
-      </td>
-      <td>{Math.floor(Math.random() * 6 + 5)} days ago</td>
-      <td>{Math.random() > 0.5 ? <Badge>Active</Badge> : <Badge color="gray">Disabled</Badge>}</td>
-      <td>
-        <Group spacing={0} position="right">
-          <ActionIcon>
-            <IconPencil size={16} stroke={1.5} />
-          </ActionIcon>
-          <Menu transition="pop" withArrow position="bottom-end">
-            <Menu.Target>
-              <ActionIcon>
-                <IconDots size={16} stroke={1.5} />
-              </ActionIcon>
-            </Menu.Target>
-            <Menu.Dropdown>
-              <Menu.Item icon={<IconMessages size={16} stroke={1.5} />}>Send message</Menu.Item>
-              <Menu.Item icon={<IconNote size={16} stroke={1.5} />}>Add note</Menu.Item>
-              <Menu.Item icon={<IconReportAnalytics size={16} stroke={1.5} />}>Analytics</Menu.Item>
-              <Menu.Item icon={<IconTrash size={16} stroke={1.5} />} color="red">
-                Terminate contract
-              </Menu.Item>
-            </Menu.Dropdown>
-          </Menu>
-        </Group>
-      </td>
-    </tr>
-  ));
-
-  return (
-    <ScrollArea>
-      <Table sx={{minWidth: 800}} verticalSpacing="sm">
-        <thead>
-          <tr>
-            <th>Employee</th>
-            <th>Role</th>
-            <th>Job title</th>
-            <th>Last active</th>
-            <th>Status</th>
-            <th className="text-right">Actions</th>
-          </tr>
-        </thead>
-        <tbody>{rows}</tbody>
-      </Table>
-    </ScrollArea>
-  );
-};
