+++ src/index.css
@@ -30,3 +30,17 @@
     background-color: #a8bbbf;
   }
 }
+
+/* for tailwind intellisense works and show detail tooltip: add custom in plugin instead of do @apply on @layer here. */
+/* ref: https://github.com/tailwindlabs/tailwindcss-intellisense/issues/227 */
+/* @layer components {
+  .card-primary {
+    @apply rounded-lg border border-gray-200 p-4
+      transition-shadow duration-150 ease-in-out
+      hover:shadow-md dark:border-gray-700 dark:bg-gray-800;
+  }
+
+  .flex-center-between {
+    @apply flex items-center justify-between gap-4;
+  }
+} */
