+++ src/pages/Setting/Setting.tsx
@@ -1,9 +1,8 @@
-import {lazy, Suspense} from 'react';
+import {Title, useMantineTheme} from '@mantine/core';
 import {useTranslation} from 'react-i18next';
-import {Skeleton, Title, useMantineTheme} from '@mantine/core';
 
+import ThemeColors from '~/features/ThemeColors';
 import {Head} from '~/outlet/Head';
-const Showcase = lazy(() => import('~/features/Demos/Showcase'));
 
 const Setting = () => {
   const {t} = useTranslation();
@@ -11,12 +10,10 @@ const Setting = () => {
 
   return (
     <>
-      <Head />
+      <Head title={t('setting.pageTitle')} />
       <Title color={theme.primaryColor}>{t('setting.pageTitle')}</Title>
 
-      <Suspense fallback={<Skeleton className="h-1/4" visible />}>
-        <Showcase />
-      </Suspense>
+      <ThemeColors />
     </>
   );
 };
